strip use dep defaults so known use flags get identified correctly
Without this, the identify_use method fails to determine the correct
flags to be enabled from the profile settings in cases where the flags
are suffixed by use dep defaults such as '(-)' because the set of known
flags with use dep default suffixes don't intersect with the set of use
flags enabled from the profile that don't have suffixes.

This should fix issues with the multilib deps from the current multilib
eclass implementation.