visibility checks: add suppression for memory exhaustion.
Given an atom like the following:

>=media-video/libav-9.12[X?,abi_mips_n32(-)?,abi_mips_n64(-)?,abi_mips_o32(-)?,abi_ppc_32(-)?,abi_ppc_64(-)?,abi_s390_32(-)?,abi_s390_64(-)?,abi_x86_32(-)?,abi_x86_64(-)?,abi_x86_x32(-)?,encode?,gsm?,jpeg2k?,mp3?,opus?,sdl?,speex?,theora?,threads?,truetype?,vaapi?,vdpau?,x264?]

current pkgcore implementation will render that into a tree of conditionals.
This results in 2^25th possible branches- exhausting memory.  Fixing this
in pkgcore is trickier than it sounds; as pcheck is aware of the problem
and will bypass those scenarios for anything over 2^16th.  This still can
result in large memory usage, but is a usable workaround until pkgcore
is properly fixed.