addons: large reworking of the profiles addon
The options --profile-disable-dev, --profile-disable-exp, and
--profile-disable-profiles-desc, --enable-profiles, and
--disable-profiles, and --disable-arches have been removed.

Now only -a/--arches and -p/--profiles options exist, but they take
comma separated lists of enabled and disabled targets. To disable
targets prefix them with '-'. Note that to start an argument list with a
disabled target an equals sign must be used, e.g. -a=-arm, otherwise
argparse thinks the disabled target is another option.

If no profiles are explicitly enabled via --profiles, then all profiles
in the targeted repo will be added for scanning minus any that are
explicitly disabled.

To replace the --profile-disable-dev and --profile-disable-exp options,
the keywords stable, dev, and exp can be used as arguments to
--profiles, e.g. to scan all profiles except development and
experimental profiles use the following command:

pkgcheck --profiles=-dev,-exp

or another way to achieve the same objective of only scanning stable
profiles:

pkgcheck --profiles stable

Also, --profile-disable-deprecated to has been renamed to
--profiles-disable-deprecated and --profile-base to --profiles-base for
consistency.