metadata_checks: Check for GLEP 73 syntax compliance
Check for REQUIRED_USE constraints that do not match the syntax
limitations imposed within GLEP 73. While the limitations are related to
the GLEP, they are generically useful to find unnecessarily complex
and unreadable constructs.

The reported constructs include all kinds of groups recursively nested
inside ||/??/^^, empty ||/??/^^ and meaningless all-of groups.