Optimize the DuplicateFiles check to checksum only matching size
Add an additional grouping by size for the DuplicateFiles check,
and checksum the files only if they have matching size. This should
save us from having to checksum all the files in the repository
on every CI run. Given that we already stat() the files, checking
for matching size has much smaller cost.