pkgcheck: -r/--repo now allows external repos
pkgcheck can now be pointed at unconfigured, external repos that aren't
in the system's repos.conf or similar.

Do note that if they are large repos without metadata, e.g. a git
checkout of the gentoo tree, it would be wise to generate the metadata
with `pmaint regen` before running scans across large cross-sections of
the repo, otherwise pkgcheck will generate metadata on the fly slowing
the process down considerably.