repo_metadata: Split Manifest conflict check out of ManifestReport
Split the Manifest checks into package-scope ManifestReport
and repo-scope ManifestConflictCheck.  This ensures that the former
is normally run when pkgcheck is called with a single package, while
the latter is run on whole repository where it can efficiently detect
checksum conflicts.  This is useful for Gentoo CI since it runs pkgcheck
per-category, and needs to run global checks separately.

Fixes #88.
Fixes #86.