BadCommandsCheck: drop check for PortageInternals
The last abuser in the gentoo repo was glibc (prepallstrip) which was fixed
recently. I don't think more should slip in easily as repoman has a
check for this and I don't see the point of keeping this around longterm
since no outfacing documentation should mention any of the prep*
functions for public use anymore.