avoid implicit namespace package imports when running
An example where this occurs is when running any of the network checks
inside the app-arch category dir in the gentoo repo. When importing the
requests module, urllib3 conditionally tries to import brotli and
imports the app-arch/brotli package directory as an implicit namespace
package even when the actual python module is missing on the system.
This causes the code to break when it tries to use the namespace as the
regular module.

Removing the current working directory from the import path list avoids
this issue.