make PerlCheck an optional check
Most users probably don't install the required perl module dep so having
this check in the default set adds an extra second or so of runtime to
the default scan run.

Those that want to enable this should just add the check to the default
set to run via `pkgcheck scan --checks=+PerlCheck` or similar.