Pipeline: store error result objs for --exit instead of a count
In theory, this could cause memory usage issues for repos with a lot of
problems. If that's ever a big issue, error objects could be written to
an external cache instead.

In the meantime, this will help support better output for CI purposes in
that results triggering a failure status can be shown in an elevated
context after a full repo scan.