Add check for ebegin-eend calls being properly paired
It was suggested to add QA warnings in Portage for calls to ebegin
missing a corresponding call to eend, and vice versa. Subsequently, it
was suggested that such lints belong in pkgcheck.

The idea is to use tree-sitter-bash to get a decent approximation for
calls to ebegin with an eend missing, and calls to eend where ebegin
hasn't been called.

Bug: https://bugs.gentoo.org/835823
Bug: https://bugs.gentoo.org/835824