add a check for remote-id validity
Add an auxiliary check for invalid remote-id values.  At the moment,
it primarily checks for the correct number of path components and some
obvious mistakes.  In the future, the regular expressions can be
improved to catch invalid user/project names as well.

This includes tests for common valid and invalid values for the current
set of remote-ids.  In addition to that, PkgMetadataXmlEmptyElement
check is extended to cover "empty <remote-id/>" path in the code.
Unfortunately, covering "unknown type" path is non-trivial as it
requires a remote-id type that is allowed by the XML Schema but not
covered by validation (which normally happens only when Schema is
updated).

This also requires bumping the minimal pkgcore version in order to grab
XML Schema updates (adding "osdn" remote type).