GitPkgCommitsCheck: fix files check
When using _RemovalRepo for previous git state repo, it uses
UnconfiguredTree::_pkg_filter as default pkg_filter. During the
check for various bad file, it tries to parse the ebuild, which in
turn tries to check the usage around inherited eclasses - which
doesn't exist in this repo.

In theory, the setting of masters repo as the original one should
suffice, but it doesn't in practice. For now, let's copy the eclass
content so at least this very important check is done!

Fixes: https://github.com/pkgcore/pkgcheck/issues/381
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>