addons/git: fix interaction with git's safe.directory
When using GitLog class, it runs the git commands with clean git config,
(aka it uses an empty one) which results in inability to use pkgcheck
scan on overlays (as it tries to access system's gentoo repo, which is
owned by root). If the user set safe.directory in his config, it will
still not work, as we override user's config.

Fix it by creating a temporary file, which we fill with minimal config
which disables safe.directory feature, and passing it as the config file
to the git log command.

Resolves: https://github.com/pkgcore/pkgcheck/issues/412
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>