GitPkgCommitsCheck: fix archival repo from multiple removals
When we are performing `--commits` scan over a commit range that have at
least 2 commits that drop a version from the same package, we can have a
state of the temporary "removal repo" that it expects one of the removed
ebuilds.

This is because the check is fed package after package, with the list of
all removals of this package in a list. When creating the archival repo,
we need to select the commit from which to "archive", and we select the
first commit in the list. This is not always the right commit, as it can
be the second or third commit that removes the a version. As a result we
have an archival repo that expects all version to exist, but one of them
might be missing.

By selecting the commit with the lowest commit date, we ensure the right
commit is selected, and all removed ebuilds are present.

On the same note, sometimes we have another exception, which involves a
failure to check the EAPI of the repo is supported, because of missing
`profiles` directory in archival repo, so also include it in the archive.

Resolves: https://github.com/pkgcore/pkgcheck/issues/460
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>