MissingRemoteIdCheck: fix bad suggestion for downloads.sourceforge.net
For sourceforge, the order of regexes was bad, resulting in it catching
`downloads.sourceforge.net` as project under `sourceforge.net` instead,
which is very wrong. Fix the order, and update the test to check for
this case specially.

On the same note, add support for `sourceforge.io` as a site - cause
looks like it also appears in the wild.

Resolves: https://github.com/pkgcore/pkgcheck/issues/488
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>