GitPkgCommitsCheck: fix modification check for added ebuild in pkgset
When commit range has modification for multiple versions, with one of
them modifying a newly added version in the same range, it wouldn't find
the correct base commit and fail. Fix it by grouping based on fullver.

The test is special with time.sleep, since I need the commits be in
different seconds, otherwise the sorting by time might be bad.

Resolves: https://github.com/pkgcore/pkgcheck/issues/563
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>