scan: fix unknown checkset during initial config load
The config load is performed in multiple stages:
1. parse `--config` path
2. load bundled config
3. load user & system configs
4. parse cmd args (needed for repo and targets)
5. load repo's config
6. parse full cmd args

The problem is that the checkset is parsed during stage 4, which means
it hasn't loaded the repo's config yet. While a more correct solution
would be to parse only needed args during stage 4, it wasn't simple
since the targets collect everything, including cmd args. So instead
this just patches temporarily to not fail upon unknown checksets.

Resolves: https://github.com/pkgcore/pkgcheck/issues/576
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>