compatibilty: remove reliance on repo.category_dirs
pkgcore's `category_dirs` is a fossil that shouldn't be
relied upon.  My recollection is this property was from
before `profiles/categories` existed, but it's just
a listdir() intersection against known directories.
PMS mandates categories be in `profiles/categories`, thus
that functionality breaks compatibility.

It will be removed from pkgcore in future semver breaks
for that reason.

Signed-off-by: Brian Harring <ferringb@gmail.com>