feat: add check for a category is a file
This is beyond corner case, but if someone has a non directory
in place of where a category is supposed to be, pkgcore and portage
will silently suppress this.

Pkgcore will be modified to emit a warning for this, but in the meantime,
it's a cheap check, thus add it.

Signed-off-by: Brian Harring <ferringb@gmail.com>