history: suppress revisions where content was being renamed or moved around.
The point here is to be able to look back through time on the evolution to
understand what/why things changed.  In particular, it's useful to trace
back and get a sense of the age of the code and use that as a gauge for the
potential that things changed w/out it being updated, or was working around
things from py2.4 era.

This is from a manual audit, so I didn't catch everything, and some
stuff (minor renames or things the history isn't worth linking) I left
out based on a quick scan of commit message and coarse file changes.

Signed-off-by: Brian Harring <ferringb@gmail.com>