BadPackageUpdate: fix when not run inside profiles/ dir
When we weren't inside profiles/ dir, we were running the VisibilityCheck
check. This check was collecting repo.config.updates. This one is cached,
resulting in the value not being recomputed when time comes for
PackageUpdatesCheck to run. This commit fixes this by making sure we
invalidate the cache.

Sadly the object we're using to store config is Immutable, resulting in
a very ugly hack to invalidate the cache, by using the __dict__ attribute
of the object. This is not a good practice, but it works.

Reported-by: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>