caches: support compression of cache files
profiles.pickle has become quite fat on disk, getting to 185MB in size.
More information on the source of issue can be found in the issue linked
below, but I've decided to use "zstd -T0" (with default compression level)
to compress the cache files. This should help with the size of the
cache files, and the performance hit should be negligible.

I've measured the time it takes to load the cache files before and after
this change, and the difference is nil. The time is mostly the cost of
pickle.load, and the compression/decompression is negligible in comparison.

I'm still somewhat concerned about my usage of subprocess.Popen, but I
think it's fine.

Resolves: https://github.com/pkgcore/pkgcheck/issues/735
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>