Replace p7zip with 7zip for .7z files
P7zip was a fork of 7zip from a time when 7zip was windows only.
Since then 7zip got official Linux support and work on p7zip stopped.
At this point p7zip is old and outdated and may or may not be vulnerable.
(See notice on the official Gentoo wiki)

Last year I filed a bug about this on the Gentoo bug tracker
which is currently being worked on.
I think it is somewhat counter productive to tell people to require p7zip
when there is active work to replace it.

https://wiki.gentoo.org/index.php?title=P7zip&oldid=1306559
https://bugs.gentoo.org/942397

Signed-off-by: Ophelia de Sica <GoGoOtaku@gmail.com>