SystemSetMissingKeywords: check for missing keywords in system set packages
Happened not long ago, that sys-apps/kbd was missing stable ppc and
ppc64 keywords, which caused the stage3 build to fail for 3 months.
This is something we can check easily, so let's do it.

This commit adds a new check to the RepoProfilesCheck class that
verifies that all packages in the system set have matching keywords
to the arches' profiles system set.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>