refactor: Fix ontology of reporters and annotations.
The base is an ABC, thus make it one.  In parallel the consuming
generator should *not* be assumed to be reusable between contexts,
instead the context should open up that consumer each time.

This code effects that change, while eliminating _start and _finish
since overloading __enter__ and __exit__ do the same thing.  Plus it's
only shitty xml that needs that.

Beyond this, annotations are threaded in so folks writing checks will
be warned if they try reporting a non Result derivative.

Via the cleanup above, .report was removed from Reporter instances.  Instead
a context object is returned with *strictly* the signature that should be
used, and additionally has a hardwired __call__ to the send in addition
to the .report().

I'll convert all checks to use the __call__ method in later commits.
That sort of flow is both clearer, and it also is faster (__call__ doesn't
require __dict__ or __slots__ lookup in the method tables- it's a hardwired
field in PyObject).

Finally, make this all immutable to keep folks from poking at shit they
shouldn't.

Signed-off-by: Brian Harring <ferringb@gmail.com>