feat: test_pkg_scan: output debugging info, allow custom hooks
First: output the source path of the expected result since it's pain
trying to debug this if you're an external user.  This states exactly
what file caused it.  For example:

Second: add the ability to use python code to filter out results.  This
can be used in parallel to the normal mechanism, and there are checks
to try and block accidental suppression across the testdata.
Custom handlers can be added as /handler.py with a 'handler' functor
with the signature of typing.Callable[[Result], bool].

Third: fix https://github.com/pkgcore/pkgcheck/issues/760 since it's
causative.  Between bash 5.2 and 5.3 the error output improved,
can't just store a literal error json.  Instead a custom bit
of python can be written to do the enforcement, which is added
here.

Fourth: add as much sanity checks against the handlers as I can think
of.  Being it's python code, addition of a random new handler can lead
to missing/unknowns- thus wire things in a way to try maximally to
identify the offender so debugging failures like this are less of a
heavy lift.

Close: pkgcore/pkgcheck#760
Signed-off-by: Brian Harring <ferringb@gmail.com>